Move getAllThemes and getThemeNameAsString to common-rendering
