+++ common-rendering/src/fixtures/article.ts
